jobs: add writeFirstLegacyPayload and writeFirstLegacyProgress internal apis
This patch adds new job info api calls which callers in the jobs package can
use to write the first job info keys. These 2 calls reduce the number of sql
statements executed on the first info key write from 2 to 1.

Epic: none

Release note: none

+++ pkg/jobs/job_info_storage.go
@@ -441,8 +441,21 @@ func (i InfoStorage) get(ctx context.Context, opName, infoKey string) ([]byte, b
 	return []byte(*value), true, nil
 }
 
-func (i InfoStorage) write(ctx context.Context, infoKey string, value []byte) error {
+func (i InfoStorage) write(
+	ctx context.Context, infoKey string, value []byte, firstWrite bool,
+) error {
 	return i.doWrite(ctx, func(ctx context.Context, j *Job, txn isql.Txn) error {
+		if firstWrite {
+			// firstWrite implies that the value is being written for the first time,
+			// so no need to update it.
+			_, err := txn.ExecEx(
+				ctx, "write-job-info-insert", txn.KV(),
+				sessiondata.NodeUserSessionDataOverride,
+				`INSERT INTO system.job_info (job_id, info_key, written, value) VALUES ($1, $2, now(), $3)`,
+				j.ID(), infoKey, value,
+			)
+			return err
+		}
 
 		deleteQuery := "DELETE FROM system.job_info WHERE job_id = $1 AND info_key::string = $2"
 		if value == nil {
@@ -597,12 +610,22 @@ func (i InfoStorage) Write(ctx context.Context, infoKey string, value []byte) er
 	if value == nil {
 		return errors.AssertionFailedf("missing value (infoKey %q)", infoKey)
 	}
-	return i.write(ctx, infoKey, value)
+	return i.write(ctx, infoKey, value, false /* firstWrite */)
+}
+
+// WriteFirstKey writes the provided value to an info record for the provided
+// jobID and infoKey via a blind insert. It is the caller's responsibility to
+// ensure that the info record does not already exist.
+func (i InfoStorage) WriteFirstKey(ctx context.Context, infoKey string, value []byte) error {
+	if value == nil {
+		return errors.AssertionFailedf("missing value (infoKey %q)", infoKey)
+	}
+	return i.write(ctx, infoKey, value, true /* firstWrite */)
 }
 
 // Delete removes the info record for the provided infoKey.
 func (i InfoStorage) Delete(ctx context.Context, infoKey string) error {
-	return i.write(ctx, infoKey, nil /* value */)
+	return i.write(ctx, infoKey, nil /* value */, false /* firstWrite */)
 }
 
 // DeleteRange removes the info records between the provided
@@ -717,4 +740,18 @@ func (i InfoStorage) GetLegacyProgress(ctx context.Context, opName string) ([]by
 // WriteLegacyProgress writes the job's Progress to the system.job_info table.
 func (i InfoStorage) WriteLegacyProgress(ctx context.Context, progress []byte) error {
 	return i.Write(ctx, LegacyProgressKey, progress)
+}
+
+// writeFirstLegacyProgress writes the job's Progress to the system.job_info table
+// via a blind insert. It is the caller's responsibility to ensure that the
+// info record does not already exist.
+func (i InfoStorage) writeFirstLegacyProgress(ctx context.Context, progress []byte) error {
+	return i.WriteFirstKey(ctx, LegacyProgressKey, progress)
+}
+
+// writeFirstLegacyPayload writes the job's Payload to the system.job_info table
+// via a blind insert. It is the caller's responsibility to ensure that the
+// info record does not already exist.
+func (i InfoStorage) writeFirstLegacyPayload(ctx context.Context, payload []byte) error {
+	return i.WriteFirstKey(ctx, LegacyPayloadKey, payload)
 }
